fix compositor behavior to work better w/ dups
The previous logic for how the compositor dealt with and stored
duplicate frames was kind of mind-breaking, so it's been reworked.
The compositor layer now assumes that when it's called, it's now
time to composite the frame that it was called with, and should move
time forward appropriately. It's up to the caller to make sure it
gets called at an appropriate time.

This means frames in both the presented frame list and the output
frame list should correctly show when duplicate output frames are
happening.