initial attempt at somewhat pluggable handlers
Creates FrameStream, GameCapture, and Compositor protocols. The
protocols are then subtyped to create the PresentMonCapture,
StockGameCapture, and StockOBSCompositor classes. Right now,
there's no way to select which get used -- the stock ones
there are the only ones that currently exist.

In the log run I'd like to move those handlers into separate
files, and provide command line options or similar to be able
to select which ones get used. That will (hopefully?) happen
in a future commit.